implemented pagination

+++ src/main/java/com/docappointment/iam/services/impl/DoctorServiceImpl.java
@@ -1,6 +1,8 @@
 package com.docappointment.iam.services.impl;
 
+import com.docappointment.iam.dto.PaginatedDoctorsDTO;
 import com.docappointment.iam.entities.PatientDetails;
+import com.docappointment.iam.exceptions.ResourceNotFoundException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.docappointment.iam.dao.DoctorDetailsDao;
 import com.docappointment.iam.dao.UserDao;
@@ -13,6 +15,10 @@ import com.docappointment.iam.enums.Role;
 import com.docappointment.iam.services.DoctorService;
 import com.docappointment.iam.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -51,24 +57,30 @@ public class DoctorServiceImpl implements DoctorService {
         Optional<User> user = userDao.findById(doctorId);
         Optional<DoctorDetails> doctorDetails = doctorDetailsDao.findById(doctorId);
 
-        if (user.isPresent()) {
+        if (user.isPresent() && user.get().getRole().equals(Role.DOCTOR)) {
             DoctorDetails doctorDetails1 = doctorDetails.orElseGet(DoctorDetails::new);
             return mergeUserAndDetails(user.get(), doctorDetails1);
+        } else {
+            throw new ResourceNotFoundException("Invalid doctor id");
         }
-
-        return null;
     }
 
     @Override
-    public List<DoctorDTO> getAllDoctors() {
+    public PaginatedDoctorsDTO getAllDoctors(int pageNumber, int pageSize) {
         List<DoctorDTO> doctors = new ArrayList<>();
-        List<User> users = userDao.findAllByRole(Role.DOCTOR);
+
+        Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(Sort.Order.asc("userId")));
+        Page<User> paginatedUsers = userDao.findByRole(Role.DOCTOR, pageable);
+        List<User> users = paginatedUsers.getContent();
+
         for (User user: users) {
             Optional<DoctorDetails> doctorDetails = doctorDetailsDao.findById(user.getUserId());
             doctorDetails.ifPresent(details -> doctors.add(mergeUserAndDetails(user, details)));
         }
 
-        return doctors;
+        PaginatedDoctorsDTO response = convertPaginatedDataToDTO(paginatedUsers);
+        response.setDoctors(doctors);
+        return response;
     }
 
     private DoctorDTO mergeUserAndDetails(User user, DoctorDetails doctorDetails) {
@@ -85,11 +97,18 @@ public class DoctorServiceImpl implements DoctorService {
 
     @Override
     public String deleteDoctor(int id) {
+        Optional<User> user = userDao.findById(id);
+        Optional<DoctorDetails> doctorDetails = doctorDetailsDao.findById(id);
+
+        if (user.isEmpty() || doctorDetails.isEmpty()) {
+            throw new ResourceNotFoundException("Invalid doctor id");
+        }
+
         try {
             userDao.deleteById(id);
             doctorDetailsDao.deleteById(id);
         } catch (Exception e) {
-            return "Failed to delete doctor";
+            throw new RuntimeException("Failed to delete doctor");
         }
 
         return "Deleted successfully";
@@ -103,12 +122,29 @@ public class DoctorServiceImpl implements DoctorService {
             Optional<User> user = userDao.findById(id);
 
             DoctorDetails doctorDetails = objectMapper.convertValue(doctorDTO, DoctorDetails.class);
-
+            doctorDetails.setUserId(user.get().getUserId());
             DoctorDetails savedDoctorDetails = doctorDetailsDao.save(doctorDetails);
             if (user.isPresent())
                 return mergeUserAndDetails(user.get(), savedDoctorDetails);
+        } else {
+            throw new ResourceNotFoundException("Invalid doctor id");
         }
 
         return null;
+    }
+
+    private PaginatedDoctorsDTO convertPaginatedDataToDTO(Page<User> paginatedData) {
+        PaginatedDoctorsDTO dto = new PaginatedDoctorsDTO();
+        dto.setPageNumber(paginatedData.getPageable().getPageNumber());
+        dto.setPageSize(paginatedData.getPageable().getPageSize());
+        dto.setTotalPages(paginatedData.getTotalPages());
+        dto.setTotalDoctors((int) paginatedData.getTotalElements());
+
+        return dto;
+    }
+
+    @Override
+    public List<String> getAllSpecializations() {
+        return doctorDetailsDao.findDistinctSpecializations();
     }
 }
